+++ myqueue/info.py
@@ -2,8 +2,12 @@ import json
 import os
 import textwrap
 from pathlib import Path
+from typing import Generator
 
 from myqueue import __version__
+from myqueue.config import Configuration
+from myqueue.pretty import pprint
+from myqueue.progress import Spinner
 from myqueue.queue import Queue
 from myqueue.selection import Selection
 
@@ -36,3 +40,40 @@ def info(queue: Queue, id: int = None) -> None:
             print('v' * N)
             print(err)
             print('^' * N)
+
+
+def info_all(start: Path) -> None:
+    """Write information about all .myqueue folders."""
+    dev = start.stat().st_dev
+    spinner = Spinner()
+    for path in scan(start, dev, spinner):
+        spinner.reset()
+        print(f'{path}:\n  ', end='')
+        try:
+            config = Configuration.read(path)
+        except FileNotFoundError as ex:
+            print(ex)
+            continue
+        with Queue(config, need_lock=False) as queue:
+            queue._read()
+            pprint(queue.tasks, short=True)
+    spinner.reset()
+    print(' ')
+
+
+def scan(path: Path,
+         dev: int,
+         spinner: Spinner) -> Generator[Path, None, None]:
+    """Scan for .myqueue folders.
+
+    Only yield paths on same filesystem (dev).
+    """
+    with os.scandir(path) as entries:
+        for entry in entries:
+            spinner.spin()
+            if entry.is_dir(follow_symlinks=False):
+                if entry.name == '.myqueue':
+                    yield path / entry.name
+                elif (not entry.name.startswith(('.', '_')) and
+                      entry.stat().st_dev == dev):
+                    yield from scan(path / entry.name, dev, spinner)