+++ myqueue/progress.py
@@ -1,4 +1,5 @@
 """Progress-bar."""
+import os
 import sys
 from typing import Iterator
 
@@ -50,16 +51,50 @@ def _progress_bar(length: int,
         yield n
 
 
-if __name__ == '__main__':
+class Spinner:
+    def __init__(self) -> None:
+        """Print the characters .oOo.oOo.oOo.oOo and so on."""
+        if sys.stdout.isatty():
+            self.fd = sys.stdout
+        else:
+            self.fd = open(os.devnull, 'w')
+        self.n = 0
+
+    def spin(self) -> None:
+        """Spin the spinner."""
+        N = 500
+        if self.n % N == 0:
+            self.fd.write('\r' + '.oOo'[(self.n // N) % 4])
+            self.fd.flush()
+        self.n += 1
+
+    def reset(self) -> None:
+        """Place cursor at the beginning of the line."""
+        self.n = 0
+        self.fd.write('\r')
+
+
+def main(x: float = 1) -> None:
+    """Demo."""
     from time import sleep
     pb = progress_bar(500, 'Test 1:')
     for _ in range(500):
-        sleep(0.002)
+        sleep(0.002 * x)
         next(pb)
     pb = progress_bar(500, 'Test 2:', 0)
     for _ in range(500):
         next(pb)
     pb = progress_bar(5, 'Test 3:')
     for _ in range(5):
-        sleep(2.5)
+        sleep(0.5 * x)
         next(pb)
+    spinner = Spinner()
+    for _ in range(10000):
+        spinner.spin()
+        sleep(0.00005 * x)
+    spinner.reset()
+    print('Done!')
+
+
+if __name__ == '__main__':
+    main()